feat(cli)!: remove the top-level `agents hosts` command group
`agents devices` is now the sole user-facing fleet registry. The registry
subcommands map to devices (list/add/check/remove -> agents devices
list/add/status/rm). The two dispatched-task commands with no devices
equivalent moved under devices: `agents hosts ps` -> `agents devices ps`
and `agents hosts stop <id>` -> `agents devices stop <id>` (alias kill).
`agents hosts logs <id>` is dropped because `agents logs <id>` already
views and follows host-dispatch task logs.

The dispatch fabric under lib/hosts/* is unchanged: `agents run
--host/--device`, teams remote placement, and cloud `--provider host`
all still work. Only the command surface is removed.

- Delete src/commands/hosts.ts; unregister loadHosts from the command
  registry (so `agents hosts` is now an unknown command).
- Add `agents devices ps` (reconciles running records, --json) and
  `agents devices stop <id>` in src/commands/ssh.ts.
- Rewrite user-facing strings that pointed at the removed command
  (lib/hosts/*, exec, message, teams, cloud) to the devices/logs
  equivalents.
- Regenerate docs/command-index.{md,json} + command-reference.html and
  update the prose docs (README, hosts.md, concepts, QUICKSTART,
  specifications, ssh-transport, terminal-engine, schema).
- CHANGELOG fragment under apps/cli/.changelog/next/.
