+++ apps/cli/src/bootstrap.ts
@@ -1288,8 +1288,8 @@ try {
       process.exit(1);
     }
     // A --host targeting a password-auth device throws this from resolveHost.
-    // It carries an actionable message (switch to key auth / enroll as a host);
-    // handling it here covers every resolveHost caller (run, hosts check/rm,
+    // It carries an actionable message (switch to key auth);
+    // handling it here covers every resolveHost caller (run, teams,
     // secrets --host) at the source instead of a catch at each call site.
     if (err.name === 'DeviceOffloadUnsupportedError') {
       console.error(err.message);
