--- apps/cli/src/commands/hosts.ts
@@ -1,358 +1,0 @@
-/**
- * `agents hosts` — register and inspect agent hosts (machines you offload runs to).
- *
- * The registry is a thin overlay: ssh-config hosts are dispatchable with zero
- * registration (connection details stay in ~/.ssh/config); enrollment only adds
- * capability metadata or bootstraps agents-cli, plus inline (non-ssh-config)
- * hosts. Dispatch itself is `agents run --host <name>` (see commands/exec.ts).
- */
-
-import type { Command } from 'commander';
-import chalk from 'chalk';
-import { terminalWidth, truncateToWidth, stringWidth } from '../lib/session/width.js';
-import { checkbox, confirm } from '@inquirer/prompts';
-import { isInteractiveTerminal } from './utils.js';
-import { assertValidSshTarget } from '../lib/ssh-exec.js';
-import { getProvider, listAllHosts, resolveHost } from '../lib/hosts/registry.js';
-import { getDevice } from '../lib/devices/registry.js';
-import { hostIdentityArgs, sshTargetFor, type Host } from '../lib/hosts/types.js';
-import { listSshConfigHosts, listKnownHosts, isSshConfigHost } from '../lib/hosts/ssh-config.js';
-import {
-  probeHost,
-  remoteAgentsVersion,
-  bootstrapAgentsCli,
-  localCliVersion,
-} from '../lib/hosts/ready.js';
-import { resolveRemoteOsSync } from '../lib/hosts/remote-os.js';
-import { listTasks, loadTask, findTaskByName, findTaskBySessionId, resolveTaskRef } from '../lib/hosts/tasks.js';
-import { reconcileRunningTasks } from '../lib/hosts/reconcile.js';
-import { showHostTaskLog } from '../lib/hosts/logs.js';
-import { stopDispatchedTask } from '../lib/hosts/dispatch.js';
-
-interface AddOptions { cap?: string[]; os?: string; enroll?: boolean; }
-
-/** Parse `user@host` or `host` into its pieces. */
-function parseTarget(target: string): { address: string; user?: string } {
-  const at = target.indexOf('@');
-  if (at === -1) return { address: target };
-  return { user: target.slice(0, at), address: target.slice(at + 1) };
-}
-
-/**
- * Bootstrap/verify agents-cli on a freshly-enrolled host (best-effort, prompts).
- * Callers that just probed the host pass the result through to avoid a second,
- * redundant `uname` round-trip; otherwise we probe here.
- */
-async function maybeBootstrap(
-  target: string,
-  hostName: string,
-  probe: { reachable: boolean; os?: string } = probeHost(target, resolveRemoteOsSync(hostName)),
-): Promise<void> {
-  if (!probe.reachable) {
-    console.log(chalk.yellow(`  Not reachable over SSH yet — skipping bootstrap. Fix key auth, then: agents hosts check ${hostName}`));
-    return;
-  }
-  // Prefer the OS the probe just observed; fall back to the registry hint.
-  const os = probe.os ?? resolveRemoteOsSync(hostName);
-  const remoteVer = remoteAgentsVersion(target, os);
-  const localVer = localCliVersion();
-  if (!remoteVer) {
-    // Non-TTY: this is a best-effort bootstrap prompt, so report and return
-    // instead of hanging on confirm() in a headless/piped shell.
-    if (!isInteractiveTerminal()) {
-      console.log(chalk.gray(`  agents-cli not found on ${hostName} — install it there, then: agents hosts check ${hostName}`));
-      return;
-    }
-    const ok = await confirm({ message: `  agents-cli not found on ${hostName}. Install ${localVer ? `v${localVer}` : 'latest'} now?`, default: true });
-    if (ok) {
-      console.log(chalk.gray('  Installing agents-cli on the host…'));
-      const r = bootstrapAgentsCli(target, localVer, os);
-      console.log(r.ok ? chalk.green('  Installed.') : chalk.red(`  Install failed:\n${r.output}`));
-    }
-    return;
-  }
-  const remoteClean = remoteVer.replace(/^v/, '');
-  if (localVer && remoteClean !== localVer) {
-    if (!isInteractiveTerminal()) {
-      console.log(chalk.gray(`  ${hostName} has agents-cli ${remoteClean}, you have ${localVer} — versions differ; upgrade the host to match when convenient.`));
-      return;
-    }
-    const ok = await confirm({ message: `  ${hostName} has agents-cli ${remoteClean}, you have ${localVer}. Upgrade to match?`, default: false });
-    if (ok) {
-      const r = bootstrapAgentsCli(target, localVer, os);
-      console.log(r.ok ? chalk.green('  Upgraded.') : chalk.red(`  Upgrade failed:\n${r.output}`));
-    }
-  }
-}
-
-async function registerHost(spec: Host): Promise<void> {
-  const provider = getProvider('local');
-  await provider.register!(spec);
-}
-
-async function doAdd(name: string | undefined, target: string | undefined, opts: AddOptions): Promise<void> {
-  // No name + no target → interactive scan of ssh sources.
-  if (!name && !target) {
-    const existing = new Set((await listAllHosts()).filter((h) => h.enrolled).map((h) => h.name));
-    const candidates = [...new Set([...listSshConfigHosts(), ...listKnownHosts()])].filter((c) => !existing.has(c));
-    if (candidates.length === 0) {
-      console.log(chalk.yellow('No SSH hosts found in ~/.ssh/config or ~/.ssh/known_hosts. Add one explicitly: agents hosts add <name> <user@host>'));
-      return;
-    }
-    const picked = await checkbox({
-      message: 'Select hosts to enroll (connection details come from ~/.ssh/config)',
-      choices: candidates.map((c) => ({ value: c, name: c })),
-    });
-    for (const c of picked) {
-      // The ssh target is the candidate name itself either way: ssh resolves it
-      // for ssh-config hosts, and it's a reachable hostname for known_hosts ones.
-      const source = isSshConfigHost(c) ? 'ssh-config' : 'inline';
-      const probe = probeHost(c);
-      await registerHost({ name: c, provider: 'local', source, ...(source === 'inline' ? { address: c } : {}), os: probe.os, caps: opts.cap });
-      console.log(chalk.green(`Enrolled ${c}`) + chalk.gray(` (${source}${probe.os ? `, ${probe.os}` : ''})`));
-      if (opts.enroll !== false) await maybeBootstrap(c, c, probe);
-    }
-    return;
-  }
-
-  if (!name) {
-    console.log(chalk.red('Usage: agents hosts add <name> [user@host]'));
-    process.exitCode = 1;
-    return;
-  }
-
-  let spec: Host | undefined;
-  let sshTarget = '';
-  if (target) {
-    assertValidSshTarget(target);
-    const { address, user } = parseTarget(target);
-    spec = { name, provider: 'local', source: 'inline', address, user, caps: opts.cap, os: opts.os };
-    sshTarget = target;
-  } else if (isSshConfigHost(name)) {
-    spec = { name, provider: 'local', source: 'ssh-config', caps: opts.cap, os: opts.os };
-    sshTarget = name;
-  } else {
-    // A registered device enrolls with no target — connection details come from
-    // the device profile. The main reason to enroll one at all is the overlay
-    // metadata (capability tags); plain dispatch already works via the devices
-    // provider.
-    const device = await getDevice(name);
-    if (device && device.auth.method !== 'password') {
-      const address = device.address.dnsName ?? device.address.ip;
-      if (address) {
-        spec = {
-          name,
-          provider: 'local',
-          source: 'inline',
-          address,
-          user: device.user,
-          caps: opts.cap,
-          os: opts.os ?? (device.platform !== 'unknown' ? device.platform : undefined),
-        };
-        sshTarget = device.user ? `${device.user}@${address}` : address;
-      }
-    }
-    if (!spec) {
-      if (device?.auth.method === 'password') {
-        console.log(chalk.red(`Device "${name}" uses password auth — dispatch needs key auth. Switch it first: agents devices config ${name} ssh.auth key`));
-      } else {
-        console.log(chalk.red(`"${name}" is not in ~/.ssh/config or the devices registry. Pass a target: agents hosts add ${name} <user@host>`));
-      }
-      process.exitCode = 1;
-      return;
-    }
-  }
-
-  const probe = probeHost(sshTarget);
-  if (!spec.os && probe.os) spec.os = probe.os;
-  await registerHost(spec);
-  console.log(chalk.green(`Enrolled ${name}`) + chalk.gray(` (${spec.source}${spec.os ? `, ${spec.os}` : ''}${spec.caps?.length ? `, caps: ${spec.caps.join(',')}` : ''})`));
-  if (opts.enroll !== false) await maybeBootstrap(sshTarget, name, probe);
-}
-
-async function doList(json: boolean): Promise<void> {
-  const hosts = await listAllHosts();
-  if (json) {
-    console.log(JSON.stringify(hosts, null, 2));
-    return;
-  }
-  if (hosts.length === 0) {
-    console.log(chalk.gray('No hosts. Enroll one: agents hosts add <name> <user@host>  (or just: agents hosts add)'));
-    return;
-  }
-  console.log(chalk.bold('NAME').padEnd(20) + chalk.bold('SOURCE').padEnd(13) + chalk.bold('TARGET').padEnd(28) + chalk.bold('CAPS'));
-  for (const h of hosts) {
-    const tgt = h.source === 'ssh-config' ? chalk.gray('(ssh-config)') : `${h.user ? h.user + '@' : ''}${h.address ?? ''}`;
-    // Cap TARGET at its column so a long user@host can't shove CAPS out of alignment.
-    const tgtW = stringWidth(tgt);
-    const tgtCol = tgtW > 28 ? truncateToWidth(tgt, 28) : tgt + ' '.repeat(28 - tgtW);
-    // Devices surface their registry as SOURCE (they're synced, not enrolled here).
-    const source = h.provider === 'devices' ? 'devices' : h.source;
-    const mark = h.dispatchable === false
-      ? chalk.yellow(' ·password-auth (no dispatch)')
-      : h.enrolled ? '' : chalk.gray(' ·available');
-    console.log(h.name.padEnd(20) + source.padEnd(13) + tgtCol + (h.caps?.join(',') ?? '') + mark);
-  }
-}
-
-async function doCheck(name: string): Promise<void> {
-  const host = await resolveHost(name);
-  if (!host) {
-    console.log(chalk.red(`Unknown host "${name}". Known: ${(await listAllHosts()).map((h) => h.name).join(', ') || '(none)'}`));
-    process.exitCode = 1;
-    return;
-  }
-  const target = sshTargetFor(host);
-  const os = host.os ?? resolveRemoteOsSync(name);
-  process.stdout.write(`Probing ${chalk.cyan(name)} (${target})… `);
-  const identityArgs = hostIdentityArgs(host);
-  const probe = probeHost(target, os, identityArgs);
-  if (!probe.reachable) {
-    console.log(chalk.red('unreachable'));
-    process.exitCode = 1;
-    return;
-  }
-  console.log(chalk.green('reachable') + chalk.gray(probe.os ? ` · ${probe.os}` : ''));
-  const ver = remoteAgentsVersion(target, probe.os ?? os, identityArgs);
-  console.log(`  agents-cli: ${ver ? chalk.green(ver) : chalk.yellow('not installed')}`);
-}
-
-async function doRemove(name: string): Promise<void> {
-  const host = await resolveHost(name);
-  if (!host || !host.enrolled) {
-    console.log(chalk.yellow(`"${name}" is not enrolled (nothing to remove).`));
-    return;
-  }
-  if (host.provider === 'devices') {
-    // The devices registry owns this entry — hosts remove would silently no-op.
-    console.log(chalk.yellow(`"${name}" comes from the devices registry. Remove it there: agents devices rm ${name}`));
-    process.exitCode = 1;
-    return;
-  }
-  await getProvider('local').remove!(name);
-  console.log(chalk.green(`Removed ${name}`));
-}
-
-async function doPs(json: boolean): Promise<void> {
-  // Heal any 'running' record whose local follower died (dropped connection,
-  // laptop sleep) against the remote `.exit` before listing, so `ps` never shows
-  // a finished run stuck at 'running'.
-  const tasks = reconcileRunningTasks(listTasks());
-  if (json) {
-    console.log(JSON.stringify(tasks, null, 2));
-    return;
-  }
-  if (tasks.length === 0) {
-    console.log(chalk.gray('No host tasks yet. Dispatch one: agents run <agent> "<task>" --host <name>'));
-    return;
-  }
-  const cols = terminalWidth();
-  console.log(chalk.bold('ID').padEnd(11) + chalk.bold('NAME').padEnd(16) + chalk.bold('HOST').padEnd(16) + chalk.bold('AGENT').padEnd(10) + chalk.bold('STATUS').padEnd(11) + chalk.bold('PROMPT'));
-  for (const t of tasks) {
-    const status = t.status === 'completed' ? chalk.green(t.status) : t.status === 'failed' ? chalk.red(t.status) : chalk.yellow(t.status);
-    const nameCol = truncateToWidth(t.name ?? chalk.gray('-'), 15).padEnd(16);
-    // Prompt fills the remaining width instead of a fixed 50-char byte slice (98-char rows).
-    const promptCol = truncateToWidth(t.prompt, Math.max(12, cols - (11 + 16 + 16 + 10 + 11)));
-    console.log(t.id.padEnd(11) + nameCol + t.host.padEnd(16) + t.agent.padEnd(10) + status.padEnd(11) + promptCol);
-  }
-}
-
-async function doLogs(ref: string, follow: boolean, full: boolean): Promise<void> {
-  // Resolve the ref as a task id first, then fall back to a `--name` handle so
-  // `agents hosts logs <name>` works, not just the opaque id.
-  const id = loadTask(ref) ? ref : (findTaskByName(ref)?.id ?? findTaskBySessionId(ref)?.id ?? ref);
-  const res = await showHostTaskLog(id, follow, full);
-  if (!res.found) {
-    console.log(chalk.red(`Unknown task "${ref}".`));
-    process.exitCode = 1;
-    return;
-  }
-  if (res.exitCode !== undefined) process.exitCode = res.exitCode;
-}
-
-async function doStop(ref: string): Promise<void> {
-  // Heal first so we don't try to kill a process that already exited.
-  const current = resolveTaskRef(ref);
-  if (!current) {
-    console.log(chalk.red(`Unknown task "${ref}".`));
-    process.exitCode = 1;
-    return;
-  }
-  const task = reconcileRunningTasks([current])[0] ?? current;
-  if (task.status !== 'running') {
-    console.log(chalk.gray(`Task ${task.id} is already ${task.status}` + (task.exitCode !== undefined ? ` (exit ${task.exitCode})` : '') + '.'));
-    return;
-  }
-  try {
-    const stopped = stopDispatchedTask(task);
-    const statusColor = stopped.status === 'completed' ? chalk.green : chalk.yellow;
-    const exitNote =
-      stopped.exitCode === 143
-        ? 'exit 143 / SIGTERM'
-        : stopped.exitCode !== undefined
-          ? `exit ${stopped.exitCode}`
-          : stopped.status;
-    console.log(
-      chalk.green(`Stopped ${stopped.id}`) +
-        chalk.gray(` on ${stopped.host}`) +
-        '  ' + statusColor(stopped.status) +
-        chalk.gray(` (${exitNote})`),
-    );
-    console.log(chalk.gray(`Logs: agents hosts logs ${stopped.id}`));
-  } catch (err: any) {
-    console.error(chalk.red(err?.message ?? err));
-    process.exitCode = 1;
-  }
-}
-
-/** Register the `agents hosts` command tree. */
-export function registerHostsCommand(program: Command): void {
-  const hosts = program
-    .command('hosts')
-    .description('Register and inspect agent hosts (machines you offload runs to with `agents run --host <name>`).');
-
-  hosts
-    .command('add [name] [target]')
-    .description('Enroll a host. With no args, pick from ~/.ssh/config + known_hosts. `target` is user@host for hosts not in ssh config.')
-    .option('--cap <cap...>', 'Capability tag(s) for routing (e.g. --cap gpu)')
-    .option('--os <os>', 'Override detected OS label')
-    .option('--no-enroll', 'Register only — skip the remote agents-cli bootstrap/version check')
-    .action((name: string | undefined, target: string | undefined, opts: AddOptions) => doAdd(name, target, opts));
-
-  hosts
-    .command('list')
-    .alias('ls')
-    .description('List enrolled + ssh-config hosts (metadata only, no probing).')
-    .option('--json', 'Output JSON')
-    .action((opts: { json?: boolean }) => doList(!!opts.json));
-
-  hosts
-    .command('check <name>')
-    .description('Probe one host: reachable? agents-cli version?')
-    .action((name: string) => doCheck(name));
-
-  hosts
-    .command('remove <name>')
-    .alias('rm')
-    .description('Remove a host from the registry (does not touch ~/.ssh/config).')
-    .action((name: string) => doRemove(name));
-
-  hosts
-    .command('ps')
-    .description('List dispatched host tasks.')
-    .option('--json', 'Output JSON')
-    .action((opts: { json?: boolean }) => doPs(!!opts.json));
-
-  hosts
-    .command('logs <id>')
-    .description('Show a host task’s concise summary; --full for the raw log, -f to follow a running one.')
-    .option('-f, --follow', 'Follow live output')
-    .option('-m, --full', 'Show the full raw combined-stdout log instead of the concise summary')
-    .action((id: string, opts: { follow?: boolean; full?: boolean }) => doLogs(id, !!opts.follow, !!opts.full));
-
-  hosts
-    .command('stop <id>')
-    .alias('kill')
-    .description('Terminate a running host task from this machine (SIGTERM process group; marks failed/143).')
-    .action((id: string) => doStop(id));
-}
