+++ apps/cli/src/lib/cloud/host.ts
@@ -4,7 +4,7 @@
  * A thin adapter over the hosts subsystem (lib/hosts/*): `agents cloud run
  * --provider host --host <name>` dispatches through the SAME detached-SSH
  * launch as `agents run --host`, and the resulting task shows up in BOTH
- * `agents cloud ps` and `agents hosts ps` — one store (the host-task sidecars
+ * `agents cloud ps` and `agents devices ps` — one store (the host-task sidecars
  * under ~/.agents/.cache/hosts/), two views. No new transport, no relay: SSH
  * is the only wire, exactly like the rest of the hosts design (docs/hosts.md).
  *
@@ -97,7 +97,7 @@ export class HostCloudProvider implements CloudProvider {
       throw new MissingTargetError(
         'host',
         'No host given. Pass --host <name> (a registered host, a device, a capability tag, or user@host).',
-        'List your machines: agents hosts list  ·  register more: agents devices sync / agents hosts add',
+        'List your machines: agents devices list  ·  register more: agents devices sync',
       );
     }
     if (resolveDispatchRepos(options).length > 0) {
