+++ apps/cli/src/lib/hosts/progress.ts
@@ -282,7 +282,7 @@ export async function followHostTask(target: string, opts: FollowOptions): Promi
   for (;;) {
     const remaining = deadline - Date.now();
     if (remaining <= 0) {
-      process.stderr.write('\n[hosts] follow timed out; the run continues on the host. Reattach with: agents hosts logs ' + opts.taskId + ' -f\n');
+      process.stderr.write('\n[hosts] follow timed out; the run continues on the host. Reattach with: agents logs ' + opts.taskId + ' -f\n');
       return -1;
     }
 
@@ -316,7 +316,7 @@ export async function followHostTask(target: string, opts: FollowOptions): Promi
     // chunks were flushed as they arrived, so reconnect from the advanced offset.
     // If no bytes arrived, back off like the old idle poll.
     if (Date.now() > deadline) {
-      process.stderr.write('\n[hosts] follow timed out; the run continues on the host. Reattach with: agents hosts logs ' + opts.taskId + ' -f\n');
+      process.stderr.write('\n[hosts] follow timed out; the run continues on the host. Reattach with: agents logs ' + opts.taskId + ' -f\n');
       return -1;
     }
