+++ apps/cli/src/lib/hosts/remote-os.ts
@@ -11,7 +11,7 @@
  *      populated fleet-wide by Tailscale sync — the reliable answer for a box
  *      like `win-mini` that was discovered, not hand-enrolled.
  *   3. The enrolled `HostEntry.os` overlay in agents.yaml (the `uname` captured
- *      at `agents hosts add` time), for hosts that live only in that overlay.
+ *      when the overlay entry was written), for hosts that live only in that overlay.
  *
  * Missing/unknown from all three → `undefined`, which `remoteShellFor` maps to
  * POSIX. Kept synchronous so the sync `agents sessions --host` fan-out can use it.
