+++ apps/cli/src/lib/types.ts
@@ -1044,7 +1044,7 @@ export interface Meta {
    */
   deviceConfig?: Record<string, unknown>;
   /**
-   * Agent-host registry keyed by host name (`agents hosts`). Portable user
+   * Agent-host registry keyed by host name (the `--host`/`--device` dispatch overlay). Portable user
    * config synced with `agents repo push/pull`. For `ssh-config` hosts this is
    * just an overlay (caps/os) — the connection details stay in ~/.ssh/config and
    * are never copied. `inline` hosts carry their own address/user.
